pyproject.toml
[project]
name = "aria"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.2",
    "gitpython>=3.1.46",
    "httpx>=0.28.1",
    "langchain>=1.2.13",
    "langchain-anthropic>=1.4.0",
    "langgraph>=1.1.3",
    "neo4j>=6.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.1",
    "redis>=7.4.0",
    "unidiff>=0.7.5",
    "uvicorn>=0.42.0",
    "voyageai>=0.3.7",
]
